re-organize codes; modify test
