+++ src/test/java/org/sagebionetworks/bridge/exporter/helper/ExportHelperTest.java
@@ -4,23 +4,13 @@ import static org.mockito.Matchers.eq;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
-import static org.sagebionetworks.bridge.exporter.helper.ExportHelper.IDENTIFIER;
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertNull;
 import static org.testng.Assert.assertTrue;
 
-import java.util.List;
-import java.util.Map;
-
-import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClient;
 import com.amazonaws.services.dynamodbv2.document.Item;
 import com.amazonaws.services.dynamodbv2.document.Table;
-import com.amazonaws.services.dynamodbv2.model.AttributeValue;
-import com.amazonaws.services.dynamodbv2.model.ScanRequest;
-import com.amazonaws.services.dynamodbv2.model.ScanResult;
 import com.google.common.base.Charsets;
-import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
 import org.apache.commons.lang3.StringUtils;
 import org.joda.time.DateTime;
@@ -41,8 +31,6 @@ public class ExportHelperTest {
     private static final String UPLOAD_DATE = "2016-05-09";
     private static final String UPLOAD_START_DATE_TIME = "2016-05-09T00:00:00.000-0700";
     private static final String UPLOAD_END_DATE_TIME = "2016-05-09T23:59:59.999-0700";
-    private static final String EXPORT_TIME_TABLE_NAME = "exportTime";
-    private static final String STUDY_TABLE_NAME = "Study";
 
     private static final DateTime UPLOAD_START_DATE_TIME_OBJ = DateTime.parse(UPLOAD_START_DATE_TIME);
     private static final DateTime UPLOAD_END_DATE_TIME_OBJ = DateTime.parse(UPLOAD_END_DATE_TIME);
@@ -113,51 +101,5 @@ public class ExportHelperTest {
         endDateTime = exportHelper.getEndDateTime(request);
 
         assertNull(endDateTime);
-    }
-
-    @Test
-    public void bootstrapStudyIdsToQueryTest() {
-        // mock ddb client with scan result as last export date time
-        AmazonDynamoDBClient mockDdbClient = mock(AmazonDynamoDBClient.class);
-
-        // mock study table and study id list
-        Table mockStudyTable = mock(Table.class);
-        when(mockStudyTable.getTableName()).thenReturn(STUDY_TABLE_NAME);
-
-        List<Map<String, AttributeValue>> studyIdList = ImmutableList.of(
-                ImmutableMap.of(IDENTIFIER, new AttributeValue().withS("ddb-foo")),
-                ImmutableMap.of(IDENTIFIER, new AttributeValue().withS("ddb-bar"))
-        );
-
-        ScanResult scanResult = new ScanResult();
-        scanResult.setItems(studyIdList);
-        ScanRequest scanRequest = new ScanRequest().withTableName(STUDY_TABLE_NAME);
-        when(mockDdbClient.scan(eq(scanRequest))).thenReturn(scanResult);
-
-        ExportHelper exportHelper = new ExportHelper();
-        exportHelper.setDdbClientScan(mockDdbClient);
-        exportHelper.setDdbStudyTable(mockStudyTable);
-
-        // mock request
-        BridgeExporterRequest request;
-        List<String> studyIdsToUpdate;
-        // daily
-        request = new BridgeExporterRequest.Builder().withDate(UPLOAD_DATE_OBJ).build();
-        studyIdsToUpdate = exportHelper.bootstrapStudyIdsToQuery(request);
-
-        assertEquals(studyIdsToUpdate.size(), 2);
-        assertEquals(studyIdsToUpdate.get(0), studyIdList.get(0).get(IDENTIFIER).getS());
-        assertEquals(studyIdsToUpdate.get(1), studyIdList.get(1).get(IDENTIFIER).getS());
-
-        // with whitelist
-        request = new BridgeExporterRequest.Builder().withStartDateTime(UPLOAD_START_DATE_TIME_OBJ)
-                .withEndDateTime(UPLOAD_END_DATE_TIME_OBJ)
-                .withStudyWhitelist(ImmutableSet.of("ddb-foo"))
-                .build();
-
-        studyIdsToUpdate = exportHelper.bootstrapStudyIdsToQuery(request);
-
-        assertEquals(studyIdsToUpdate.size(), 1);
-        assertEquals(studyIdsToUpdate.get(0), "ddb-foo");
     }
 }
